pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skillio"
version = "0.1.0"
description = "AI Agent's capability discovery and assembly hub - Intent to Skill matching engine"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Skillio Team", email = "hello@skillio.dev" }
]
keywords = [
    "ai",
    "agent",
    "skill",
    "capability",
    "intent",
    "cursor",
    "claude",
    "llm"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
llm = [
    "anthropic>=0.18.0",
    "openai>=1.0.0",
]
all = [
    "skillio[dev,llm]",
]

[project.scripts]
skillio = "skillio.cli:main"

[project.urls]
Homepage = "https://skillio.dev"
Documentation = "https://skillio.dev/docs"
Repository = "https://github.com/copyboy/skillio"
Issues = "https://github.com/copyboy/skillio/issues"

[tool.hatch.build.targets.wheel]
packages = ["skillio"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
